feat : 지역목록 리스트 사이드바에 렌더링
