--- src/components/sideBar/Location.js
@@ -1,13 +1,0 @@
-// 지역 목록 container
-import React from 'react';
-import LocationList from './LocationList';
-
-const Location = () => {
-  return (
-    <div>
-      <LocationList />
-    </div>
-  );
-};
-
-export default Location;
