+++ src/components/sideBar/LocationList.js
@@ -1,11 +1,39 @@
-// 지역 목록 item - 지역 이름
+// 지역 목록 item
 import React from 'react';
 import styled from 'styled-components';
+import { useNavigate } from 'react-router-dom';
 
-const LocationList = ({ name, lat, lon, onClick }) => {
+const LocationList = ({ name, icon, temp, lat, lon }) => {
+  const navigate = useNavigate();
+
+  // 지역 리스트 클릭 시,
+  // 1. 해당 지역의 위도 경도를 localStorage에 저장
+  // 2. loading navigate
+  const addLocation = (lat, lon, name) => {
+    localStorage.setItem(
+      'SelectedLocation',
+      JSON.stringify({ lat, lon, name })
+    );
+    navigate('/', {
+      replace: false,
+      state: {
+        isFirstLoading: false,
+      },
+    });
+  };
+
   return (
-    <List onClick={() => onClick(lat, lon, name)}>
+    <List onClick={() => addLocation(lat, lon, name)}>
       <h4>{name}</h4>
+      {icon && (
+        <div className="weather">
+          <img
+            src={`http://openweathermap.org/img/wn/${icon}@2x.png`}
+            alt="weather icon"
+          />
+          {temp}°C
+        </div>
+      )}
     </List>
   );
 };
@@ -14,24 +42,27 @@ const List = styled.div`
   display: flex;
   justify-content: space-between;
   align-items: center;
+  margin: 0 8px;
+  margin-bottom: 15px;
   border: 2px solid black;
   border-radius: 15px;
   padding: 2px 20px;
-  margin: 0 8px;
-  margin-bottom: 15px;
   cursor: pointer;
 
   h4 {
     font-size: 12px;
   }
 
-  .temp {
+  .weather {
     width: 50px;
     display: flex;
-    justify-content: space-between;
+    justify-content: center;
+    align-items: center;
+    padding: 0 10px;
 
-    span {
-      font-size: 15px;
+    img {
+      width: 40px;
+      padding: 0 10px;
     }
   }
 `;
