+++ src/components/sideBar/SideBar.js
@@ -3,9 +3,9 @@ import React, { useState } from 'react';
 import styled from 'styled-components';
 import SideHeader from './SideHeader';
 import SearchBar from '../addLocation/SearchBar';
-import CurrentLocation from './CurrentLocation';
+import LocationList from './LocationList';
 
-const SideBar = ({ onClose, cityName, weather }) => {
+const SideBar = ({ onClose }) => {
   // 검색창 보여주는 상태값
   const [search, setSearch] = useState(false);
 
@@ -23,6 +23,11 @@ const SideBar = ({ onClose, cityName, weather }) => {
     }
   };
 
+  // localStorage에 저장된 지역목록 가져오기
+  const cityList = JSON.parse(localStorage.getItem('CityList'));
+
+  // 현재 위치 정보 추가 및 sorting
+
   return (
     <SidebarContainer>
       <ModalOverlay onClick={onMaskClick} />
@@ -31,7 +36,9 @@ const SideBar = ({ onClose, cityName, weather }) => {
         {search ? (
           <SearchBar hide={onClickLocationCancel} />
         ) : (
-          <CurrentLocation cityName={cityName} weather={weather} />
+          cityList.map((location, idx) => (
+            <LocationList key={idx} {...location} />
+          ))
         )}
       </ModalWrapper>
     </SidebarContainer>
